Revert "Add implementation of telemetry consent flag for kedro run (#3876)"
This reverts commit 570f66ac15df27bd3f15b64f0013fbfe17d4cdc1.
