+++ tests/framework/cli/test_cli.py
@@ -934,91 +934,3 @@ class TestRunCommand:
             " does not exist."
         )
         assert expected_output in result.output
-
-    @mark.parametrize(
-        "input",
-        ["yes", "YES", "y", "Y", "yEs"],
-    )
-    def test_run_telemetry_flag_yes(
-        self, fake_project_cli, fake_metadata, fake_session, mocker, input
-    ):
-        result = CliRunner().invoke(
-            fake_project_cli, ["run", "--telemetry", input], obj=fake_metadata
-        )
-        assert not result.exit_code
-
-        fake_session.run.assert_called_once_with(
-            tags=(),
-            runner=mocker.ANY,
-            node_names=(),
-            from_nodes=[],
-            to_nodes=[],
-            from_inputs=[],
-            to_outputs=[],
-            load_versions={},
-            pipeline_name=None,
-            namespace=None,
-        )
-
-        runner = fake_session.run.call_args_list[0][1]["runner"]
-        project_path = str(fake_metadata.project_path)
-        telemetry_file_path = Path(project_path + "/.telemetry")
-        assert isinstance(runner, SequentialRunner)
-        assert not runner._is_async
-        assert telemetry_file_path.exists()
-
-        with open(telemetry_file_path) as file:
-            file_content = file.read()
-            target_string = "consent: true"
-
-        assert target_string in file_content
-
-    @mark.parametrize(
-        "input",
-        ["no", "NO", "n", "N", "No"],
-    )
-    def test_run_telemetry_flag_no(
-        self, fake_project_cli, fake_metadata, fake_session, mocker, input
-    ):
-        result = CliRunner().invoke(
-            fake_project_cli, ["run", "--telemetry", input], obj=fake_metadata
-        )
-        assert not result.exit_code
-
-        fake_session.run.assert_called_once_with(
-            tags=(),
-            runner=mocker.ANY,
-            node_names=(),
-            from_nodes=[],
-            to_nodes=[],
-            from_inputs=[],
-            to_outputs=[],
-            load_versions={},
-            pipeline_name=None,
-            namespace=None,
-        )
-
-        runner = fake_session.run.call_args_list[0][1]["runner"]
-        project_path = str(fake_metadata.project_path)
-        telemetry_file_path = Path(project_path + "/.telemetry")
-        assert isinstance(runner, SequentialRunner)
-        assert not runner._is_async
-        assert telemetry_file_path.exists()
-
-        with open(telemetry_file_path) as file:
-            file_content = file.read()
-            target_string = "consent: false"
-
-        assert target_string in file_content
-
-    def test_run_telemetry_flag_invalid(
-        self, fake_project_cli, fake_metadata, fake_session, mocker
-    ):
-        result = CliRunner().invoke(
-            fake_project_cli, ["run", "--telemetry", "wrong"], obj=fake_metadata
-        )
-        assert result.exit_code == 1
-        assert (
-            "'wrong' is an invalid value for example pipeline or telemetry consent. It must contain only y, n, YES, or NO (case insensitive)."
-            in result.output
-        )
